@@ -5,40 +5,38 @@
   "version": "0.1.0",
   "author": "icatproject",
   "dependencies": {
-    "@emotion/core": "^10.0.6",
-    "@fortawesome/fontawesome-svg-core": "^1.2.13",
-    "@fortawesome/free-solid-svg-icons": "^5.7.0",
-    "@fortawesome/react-fontawesome": "^0.1.4",
-    "@gatsby-contrib/gatsby-plugin-elasticlunr-search": "^2.2.1",
-    "axios": "^0.18.0",
+    "@emotion/core": "^10.0.28",
+    "@fortawesome/fontawesome-svg-core": "^1.2.30",
+    "@fortawesome/free-solid-svg-icons": "^5.14.0",
+    "@fortawesome/react-fontawesome": "^0.1.11",
+    "@gatsby-contrib/gatsby-plugin-elasticlunr-search": "^2.4.2",
+    "axios": "^0.19.2",
     "elasticlunr": "^0.9.5",
-    "gatsby": "^2.0.106",
-    "gatsby-plugin-catch-links": "^2.0.9",
-    "gatsby-plugin-emotion": "^4.0.1",
-    "gatsby-plugin-manifest": "^2.0.15",
-    "gatsby-plugin-react-helmet": "^3.0.2",
-    "gatsby-plugin-sharp": "^2.0.19",
-    "gatsby-remark-copy-linked-files": "^2.0.8",
-    "gatsby-remark-images": "^3.0.1",
-    "gatsby-remark-relative-images": "^0.2.2",
-    "gatsby-source-filesystem": "^2.0.23",
-    "gatsby-transformer-remark": "^2.2.2",
-    "i": "^0.3.6",
-    "npm": "^6.7.0",
-    "prop-types": "^15.6.2",
-    "query-string": "^6.2.0",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
-    "react-helmet": "^5.2.0",
-    "remark": "^10.0.1",
-    "unist-util-visit": "^1.4.0"
+    "gatsby": "^2.24.12",
+    "gatsby-plugin-catch-links": "^2.3.11",
+    "gatsby-plugin-emotion": "^4.3.10",
+    "gatsby-plugin-manifest": "^2.4.21",
+    "gatsby-plugin-react-helmet": "^3.3.10",
+    "gatsby-plugin-sharp": "^2.6.23",
+    "gatsby-remark-copy-linked-files": "^2.3.12",
+    "gatsby-remark-images": "^3.3.22",
+    "gatsby-remark-relative-images": "^0.3.0",
+    "gatsby-source-filesystem": "^2.3.23",
+    "gatsby-transformer-remark": "^2.8.27",
+    "prop-types": "^15.7.2",
+    "query-string": "^6.13.1",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-helmet": "^6.1.0",
+    "remark": "^12.0.1",
+    "unist-util-visit": "^2.0.3"
   },
   "license": "MIT",
   "scripts": {
     "build": "gatsby build",
     "develop": "gatsby develop",
     "start": "npm run develop",
-    "format": "prettier --write \"src/**/*.js\" --write \"src/**/*.jsx\"",
+    "format": "prettier --write \"src/**/*.{js,jsx}\"",
     "lint": "eslint --ext js,jsx src --fix",
     "test": "echo \"Write tests! -> https://gatsby.app/unit-testing\"",
     "deploy": "gatsby build && gh-pages -d public -b master"
@@ -50,24 +48,22 @@
   },
   "lint-staged": {
     "*.{js,jsx}": [
-      "pretty-quick --staged",
-      "eslint --ext js,jsx src --fix",
-      "git add"
+      "eslint --fix",
+      "prettier --config .prettierrc --write"
     ]
   },
   "devDependencies": {
-    "eslint": "^5.13.0",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-config-prettier": "^4.0.0",
-    "eslint-plugin-import": "^2.16.0",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-prettier": "^3.0.1",
-    "eslint-plugin-react": "^7.12.4",
-    "gh-pages": "^2.0.1",
-    "husky": "^1.3.1",
-    "lint-staged": "^8.1.3",
-    "prettier": "^1.16.2",
-    "pretty-quick": "^1.10.0"
+    "eslint": "^7.5.0",
+    "eslint-config-airbnb": "^18.2.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-jsx-a11y": "^6.3.1",
+    "eslint-plugin-prettier": "^3.1.4",
+    "eslint-plugin-react": "^7.20.4",
+    "gh-pages": "^3.1.0",
+    "husky": "^4.2.5",
+    "lint-staged": "^10.2.11",
+    "prettier": "^2.0.5"
   },
   "repository": {
     "type": "git",
